daily-content-generator/requirements.txt
azure-functions
mongoengine>=0.24.0
google-generativeai==0.3.2
protobuf==4.23.4
google-api-core==2.15.0
googleapis-common-protos==1.62.0
azure-appconfiguration-provider>=1.2.0
